+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - `AnyVec::drain`.
 - `AnyVec` iterators.
 - `AnyVec::clone_empty`, to construct `AnyVec` of the same type.
+- `IntoIterator` implemented.
 
 ## 0.7.0
 ### Added